Hjy1030

+++ src/components/quiztype/quizCreate/AdminFillInTheBlankCreate.vue
@@ -1,29 +1,28 @@
 <template>
   <q-form class="form-container">
     <div class="title-container">
-      <q-title class="title">빈칸 채우기형</q-title>
+      <q-toolbar-title class="title">빈칸 채우기형</q-toolbar-title>
     </div>
     <div>
       <q-card>
         <!-- 과목과 챕터 선택 -->
         <q-card-section class="select-container">
           <div class="select-subject">
-            <q-label class="label-subject">과목선택</q-label>
+            <label class="label-subject">과목선택</label>
             <q-select
               class="select-box"
               v-model="subject"
-              :options="subjectOptions"
+              :options="subjectOptions.map(s => s.subject)"
               outlined
               dense
-              @update:model-value="updateDetailSubjectOptions"
             />
           </div>
           <div class="select-chapter">
-            <q-label class="label-chapter">챕터선택</q-label>
+            <label class="label-chapter">챕터선택</label>
             <q-select
               class="select-box"
               v-model="chapter"
-              :options="filteredDetailSubjectOptions.slice().reverse()"
+              :options="chapterOptions"
               outlined
               dense
             />
@@ -35,20 +34,22 @@
           <div>
             <label for="file">
               <div class="upload-button">
-                <img src="/add-image.png" alt="Upload Image" />
+                <img src="/addimg.png" alt="Upload Image" />
               </div>
             </label>
             <input type="file" id="file" @change="fileInputHandler" />
           </div>
           <div v-if="filePreview" class="previewImage-container">
             <img :src="filePreview" alt="File Preview" class="preview-image" />
-            <div class="cancel-button" @click="cancelFile">X</div>
+            <div class="cancel-button" @click="cancelFile">
+              <q-icon name="close" />
+            </div>
           </div>
         </q-card-section>
 
         <!-- 문제 입력 -->
         <q-card-section class="quiz-container">
-          <q-label class="label-quiz">질문</q-label>
+          <label class="label-quiz">질문</label>
           <q-input
             v-model="quiz"
             type="textarea"
@@ -56,7 +57,7 @@
             outlined
             dense
             placeholder="문제를 입력해주세요."
-            maxlength="300"
+            maxlength="400"
             counter
             class="input-quiz"
           /><q-icon name="help" class="help-icon1">
@@ -69,7 +70,7 @@
 
         <!-- 빈칸 입력 동적 생성 -->
         <q-card-section class="answer-container">
-          <q-label class="label-answer">빈칸 답안</q-label>
+          <label class="label-answer">빈칸 답안</label>
           <div
             v-for="(blank, index) in blankInputs"
             :key="index"
@@ -84,7 +85,7 @@
               outlined
               dense
               :placeholder="`( ${index + 1} ) 의 답안을 입력해주세요.`"
-              maxlength="300"
+              maxlength="400"
               counter
               style="flex-grow: 1"
             />
@@ -99,7 +100,7 @@
 
         <!-- 해설 입력 -->
         <q-card-section class="comment-container">
-          <q-label class="label-quiz">해설 </q-label>
+          <label class="label-quiz">해설 </label>
           <q-input
             v-model="commentary"
             type="textarea"
@@ -107,7 +108,7 @@
             outlined
             placeholder="해설을 입력하세요"
             dense
-            maxlength="300"
+            maxlength="400"
             counter
             class="input-commentary"
           />
@@ -120,24 +121,24 @@
       </q-card>
     </div>
   </q-form>
-
-  <!-- 문제 생성 성공 컴포넌트 -->
-  <SubmitQuizSuccess
-    v-if="submitQuizSuccess"
-    :submit-quiz-success="submitQuizSuccess"
-    :quiz-id="quizId"
-  />
 </template>
 
 <script setup>
 import { ref, onMounted, watch } from 'vue';
+import { useRouter } from 'vue-router';
 import {
   submitQuiz,
+  submitQuizImageTemp,
   submitQuizImage,
   deleteQuiz,
-} from 'src/services/quiz/adminQuiz.js';
-import SubmitQuizSuccess from 'src/components/quiz/SubmitQuizSuccess.vue';
-import useCategories from 'src/services/useCategories.js';
+} from 'src/services/quiz/admin/adminQuiz.js';
+import { useCategorie } from 'src/services/quiz/admin/useCategorie.js';
+import {
+  quizSuccessNotification,
+  quizErrorNotification,
+  errorNotification,
+} from 'src/services/quiz/notifications.js';
+
 //반응형 데이터
 const subject = ref('과목을 선택 해주세요.');
 const chapter = ref('챕터를 선택 해주세요.');
@@ -154,20 +155,23 @@ const goBack = () => {
 //이미지 업로드 로직
 const fileName = ref('');
 const filePreview = ref(null); // 이미지 미리보기 URL
+//이미지 업로드 로직
 const fileInputHandler = event => {
   const files = event.target && event.target.files;
   if (files && files[0]) {
     fileName.value = files[0].name;
-    // 파일 타입이 이미지인지 확인
-    if (files[0].type.startsWith('image/')) {
+    const fileType = files[0].type;
+    // 파일 타입이 JPG 또는 JPEG 인지 확인
+    if (fileType === 'image/jpeg') {
       const reader = new FileReader();
       reader.onload = e => {
         filePreview.value = e.target.result;
       };
       reader.readAsDataURL(files[0]);
     } else {
-      alert('이미지 파일만 선택할 수 있습니다.');
+      alert('JPG 파일만 선택할 수 있습니다.');
       filePreview.value = null; // 파일 미리보기 초기화
+      document.getElementById('file').value = ''; // 파일 입력 초기화
     }
   }
 };
@@ -178,26 +182,20 @@ const cancelFile = () => {
   document.getElementById('file').value = ''; // 파일 입력 초기화
 };
 
-// 과목,챕터 불러오기 로직
-const { subjectOptions, fetchCategories, getDetailSubjectsBySubject } =
-  useCategories();
-
-onMounted(fetchCategories);
-
-// 과목 선택에 따라 챕터 옵션을 업데이트하는 함수
-const filteredDetailSubjectOptions = ref([]);
-const updateDetailSubjectOptions = () => {
-  const detailSubjects = getDetailSubjectsBySubject(subject.value);
-  if (detailSubjects.length === 0) {
-    filteredDetailSubjectOptions.value = ['공백'];
-  } else {
-    filteredDetailSubjectOptions.value = detailSubjects;
+// 과목, 챕터 불러오기 로직
+const {
+  subjectOptions,
+  chapterOptions,
+  fetchSubjects,
+  selectSubject,
+  fetchChapters,
+} = useCategorie();
+onMounted(fetchSubjects);
+watch(subject, newSubject => {
+  if (newSubject) {
+    selectSubject(newSubject);
+    chapter.value = '챕터를 선택 해주세요.';
   }
-};
-watch(subject, () => {
-  // 과목이 변경될 때마다 챕터 선택 초기화
-  chapter.value = '챕터를 선택 해주세요.';
-  updateDetailSubjectOptions();
 });
 
 // 문제에 포함된 <<빈칸>>의 개수를 최대 3개로 제한하는 watch
@@ -236,11 +234,10 @@ const normalizeAnswers = blankInputs => {
   );
 };
 
-// 서버 전송 여부
-const submitQuizSuccess = ref(false);
-
 // 서버에서 받아온 퀴즈 아이디
 const quizId = ref('');
+
+const router = useRouter();
 
 // 서버에 문제 제출
 const submitQuizForm = async () => {
@@ -255,7 +252,6 @@ const submitQuizForm = async () => {
   // 입력값 검증
   let hasError = false;
   let errorMessage = '';
-
   if (subject.value === '과목을 선택 해주세요.') {
     errorMessage = '과목을 선택해 주세요.';
     hasError = true;
@@ -275,40 +271,64 @@ const submitQuizForm = async () => {
     errorMessage = '해설을 입력해 주세요.';
     hasError = true;
   }
-
   if (hasError) {
     alert(errorMessage);
     return; // 오류가 있을 경우 제출을 중단합니다.
   }
-
-  //서버에 보낼 퀴즈 데이터
-  const quizData = {
-    subject: subject.value,
-    chapter: chapter.value,
-    quizType: '5',
-    jsonContent: JSON.stringify({
-      quiz: quiz.value,
-      answer: normalizedAnswers,
-      commentary: commentary.value,
-    }),
-    hasImage: !!filePreview.value,
-  };
-  console.log('서버에 제출될 데이터:', quizData);
+  const confirmation = confirm('문제를 등록하시겠습니까? ');
+  if (!confirmation) {
+    return;
+  }
   try {
-    // 문제 데이터 서버에 제출후 반환된 퀴즈 ID 저장.
-    quizId.value = await submitQuiz(quizData);
+    //서버에 보낼 퀴즈 데이터
+    const quizData = {
+      subject: subject.value,
+      chapter: chapter.value,
+      quizType: '5',
+      jsonContent: JSON.stringify({
+        quiz: quiz.value,
+        answer: normalizedAnswers,
+        commentary: commentary.value,
+      }),
+      hasImage: !!filePreview.value,
+    };
+    //console.log('서버에 제출될 데이터:', quizData);
+    try {
+      // 이미지가 있는 경우
+      if (filePreview.value) {
+        // 이미지 임시로 제출
+        const uuid = await submitQuizImageTemp(filePreview.value);
+        // 이미지 오류 처리
+        if (!uuid) {
+          throw new Error('이미지 업로드에 실패했습니다.');
+        }
 
-    // 이미지가 있다면 이미지 서버에 제출
-    if (filePreview.value) {
-      await submitQuizImage(quizId.value, filePreview.value);
+        quizId.value = await submitQuiz(quizData);
+
+        // 이미지 진짜 제출
+        await submitQuizImage(quizId.value, uuid);
+
+        // 성공
+        quizSuccessNotification();
+        router.push(`/quizzes/${quizId.value}`);
+      } else {
+        // 이미지 없을 시 바로 제출
+        quizId.value = await submitQuiz(quizData);
+        // 성공
+        quizSuccessNotification();
+        router.push(`/quizzes/${quizId.value}`);
+      }
+    } catch (error) {
+      if (error.response?.status === 400) {
+        quizErrorNotification();
+        await deleteQuiz(quizId.value); // 퀴즈 삭제 처리
+      } else {
+        errorNotification();
+        await deleteQuiz(quizId.value); // 퀴즈 삭제 처리
+      }
     }
-
-    submitQuizSuccess.value = true; //퀴즈 제출 성공
   } catch (error) {
-    if (error.response?.status === 400) {
-      alert('오류가 발생했습니다. 다시 시도해주세요.');
-      await deleteQuiz(quizId.value); // 퀴즈 삭제 처리
-    }
+    errorNotification();
   }
 };
 </script>
@@ -336,7 +356,24 @@ const submitQuizForm = async () => {
   font-size: 2rem;
   color: #0080ff;
 }
-
+//가이드 스타일
+.guidecontainer {
+  display: flex;
+  justify-content: space-evenly; /* 중앙 정렬 */
+  align-items: center; /* 수직 중앙 정렬 */
+  width: 60px;
+  height: 60px;
+  background-color: #ebf0f1;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  border-radius: 50%;
+  cursor: pointer;
+  margin: 0 1%;
+}
+.span {
+  color: red;
+}
 // 과목 챕터 선택 스타일
 .select-container {
   display: flex;
@@ -427,6 +464,10 @@ input[type='file'] {
   text-align: center;
   line-height: 50px;
   border-radius: 50%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  cursor: pointer;
 }
 
 // 문제 입력 스타일
@@ -495,9 +536,27 @@ input[type='file'] {
 }
 .btn-back {
   background-color: rgb(213, 213, 213);
+  color: black;
+  padding: 6px 16px;
+  border-radius: 6px;
+  font-weight: 700;
+  font-size: 0.875rem;
+  line-height: 1.71429;
+  height: 40px;
+  font-family: 'Toss Product Sans';
+  box-shadow: rgba(39, 40, 40, 0.24) 0px 8px 16px 0px;
 }
 .btn-submit {
-  background-color: primary;
+  background-color: rgb(0, 154, 233);
+  color: white;
+  padding: 6px 16px;
+  border-radius: 6px;
+  font-weight: 700;
+  font-size: 0.875rem;
+  line-height: 1.71429;
+  height: 40px;
+  font-family: 'Toss Product Sans';
+  box-shadow: rgba(0, 154, 233, 0.24) 0px 8px 16px 0px;
 }
 
 .tooltip {
